pyproject.toml
[project]
name = "multi-agent-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "Ravi"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.124.0,<0.125.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "redis (>=7.1.0,<8.0.0)",
    "sqlalchemy (>=2.0.0)",
    "psycopg2-binary (>=2.9.0)",
    "asyncpg (>=0.29.0)",
    "bcrypt (>=4.0.0)",
    "passlib[bcrypt] (>=1.7.4)",
    "python-jose[cryptography] (>=3.3.0)",
    "python-dotenv (>=1.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]

[tool.poetry]
packages = [
    { include = "app" }
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
